Change file/module names to conform to Rails standards
