+++ app/lib/celery_script/store_celery.rb
@@ -4,6 +4,7 @@
 # mechanism (SQL).
 # To get around the limitation, we must convert sequence JSON from canonical to
 # flat forms. `StoreCelery` handles the conversion and storage of CS Nodes.
+module CeleryScript
 class StoreCelery < Mutations::Command
   required do
     model :sequence, class: Sequence
@@ -14,7 +15,7 @@ class StoreCelery < Mutations::Command
       sequence.primary_nodes.destroy_all
       sequence.edge_nodes.destroy_all
       first_pass  = FirstPass.run!(nodes:  flat_ir, sequence: sequence)
-      second_pass = SecondPass.run!(nodes: first_pass)
+      second_pass = CeleryScript::SecondPass.run!(nodes: first_pass)
       second_pass.map(&:save!)
     end
   end
@@ -24,4 +25,5 @@ private
   def flat_ir
     @flat_ir ||= Slicer.new.run!(sequence.as_json.deep_symbolize_keys)
   end
+end
 end
